ajustes fluxo

+++ src/Web/Controller/ContatoController.cs
@@ -1,30 +1,29 @@
-﻿namespace tests_.src.Web.Controller
-{
-    using global::tests_.src.Application.Interface.ContatoInterface;
-    using global::tests_.src.Domain.Entities;
-    using Microsoft.AspNetCore.Mvc;
-    using System.Collections.Generic;
+using LigChat.Backend.Domain.Entities;
+using Microsoft.AspNetCore.Mvc;
+using tests_.src.Application.Interface.ContatoInterface;
 
+namespace tests_.src.Web.Controller
+{
     [ApiController]
     [Route("api/[controller]")]
-    public class ContatosController : ControllerBase
+    public class ContatoController : ControllerBase
     {
         private readonly IContatoService _contatoService;
 
-        public ContatosController(IContatoService contatoService)
+        public ContatoController(IContatoService contatoService)
         {
             _contatoService = contatoService;
         }
 
         [HttpGet]
-        public ActionResult<IEnumerable<Contato>> GetContatos(int sectorId)
+        public ActionResult<IEnumerable<Contact>> GetContatos([FromQuery] int sectorId)
         {
             var contatos = _contatoService.GetAll(sectorId);
             return Ok(contatos);
         }
 
         [HttpGet("{id}")]
-        public ActionResult<Contato> GetContato(int id)
+        public ActionResult<Contact> GetContato(int id)
         {
             var contato = _contatoService.GetById(id);
             if (contato == null)
@@ -35,16 +34,26 @@
         }
 
         [HttpPost]
-        public ActionResult<Contato> CreateContato(Contato contato)
+        public ActionResult<Contact> CreateContato([FromBody] Contact contato)
         {
+            if (!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
+
             var createdContato = _contatoService.Create(contato);
             return CreatedAtAction(nameof(GetContato), new { id = createdContato.Id }, createdContato);
         }
 
         [HttpPut("{id}")]
-        public IActionResult UpdateContato(int id, Contato contato)
+        public IActionResult UpdateContato(int id, [FromBody] Contact contato)
         {
-            if (id == null)
+            if (!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
+
+            if ((int)contato.Id != id)
             {
                 return BadRequest();
             }
@@ -56,9 +65,14 @@
         [HttpDelete("{id}")]
         public IActionResult DeleteContato(int id)
         {
+            var contato = _contatoService.GetById(id);
+            if (contato == null)
+            {
+                return NotFound();
+            }
+
             _contatoService.Delete(id);
             return NoContent();
         }
     }
-
-}
+}
